Extract JsonSchemaErrorMapper to lower interceptor complexity
PHPMD's class complexity for JsonSchemaInterceptor hit 52 after the
previous inline-split. Move the validator-row -> DTO normalization to
its own internal `JsonSchemaErrorMapper` class. The interceptor now
just delegates per row, which keeps it below the complexity threshold
and gives the mapping logic an obvious home (single responsibility).

No behaviour change. Mapper is stateless; instantiated inline in
`throwJsonSchemaException` so no DI module update is needed.
